% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simulated data}
\format{
A data frame with 1000 rows and 9 variables:
}
\source{
simulated data
}
\usage{
exampleData
}
\description{
A dataset containing 1000 raws and 9 columns
}
\keyword{datasets}
