% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm3grps.R
\name{psm3grps}
\alias{psm3grps}
\title{Extending the Classical Propensity Score Matching Algorithm for Three-Group Datasets}
\usage{
psm3grps(dt, group_var = "treat", times = 1, id = "id")
}
\arguments{
\item{dt}{A data frame that includes the grouping variable along with other variables requiring balancing.}

\item{group_var}{A character string specifying the grouping variable. If not provided, it defaults to "treat".}

\item{times}{A numeric value indicating the number of iterations for the algorithm. If the outcome is not adequately balanced, executing the algorithm multiple times (more than once) is recommended.}

\item{id}{a character string for variable name of patient id in the data frame.}
}
\value{
a data frame of balanced variables,or a list of data frames if the times > 1
}
\description{
The psm3grps R package enhances the classic propensity score matching (PSM) for datasets with three groups, aiding in balancing and comparing these groups in observational studies.
}
\examples{

rm(list=ls())
data("exampleData",package="psm3grps")

# check unbalanced dataset
compareGroups::descrTable(treat~.,method = 4,data = exampleData)

# run the code
df.psm<-psm3grps(dataset)

# check the result
compareGroups::descrTable(treat~.,method = 4,data = df.psm)



}
