% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm3grps_v2.R
\name{psm3grps_v2}
\alias{psm3grps_v2}
\title{Extending the Classical Propensity Score Matching Algorithm for Three-Group Datasets}
\usage{
psm3grps_v2(dt, group_var = "treat", id = "id")
}
\arguments{
\item{dt}{A data frame that includes the grouping variable along with other variables requiring balancing.}

\item{group_var}{A character string specifying the grouping variable. If not provided, it defaults to "treat".}

\item{id}{a character string for variable name of patient id in the data frame.}
}
\value{
a data frame of balanced variables.
}
\description{
This is the foundational function for the primary function "psm3grps" within the package.
}
\examples{
df.final<-psm3grps_v2(df)
}
